# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WireGuard Mesh Manager API",
    "description": "Backend API for managing WireGuard VPN networks and devices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/networks/": {
      "get": {
        "tags": [
          "networks",
          "networks"
        ],
        "summary": "List Networks",
        "description": "List all WireGuard networks.",
        "operationId": "list_networks_api_networks__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WireGuardNetworkResponse"
                  },
                  "type": "array",
                  "title": "Response List Networks Api Networks  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "networks",
          "networks"
        ],
        "summary": "Create Network",
        "description": "Create a new WireGuard network.",
        "operationId": "create_network_api_networks__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WireGuardNetworkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireGuardNetworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/networks/{network_id}": {
      "get": {
        "tags": [
          "networks",
          "networks"
        ],
        "summary": "Get Network",
        "description": "Get a specific WireGuard network by ID.",
        "operationId": "get_network_api_networks__network_id__get",
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Network Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireGuardNetworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "networks",
          "networks"
        ],
        "summary": "Update Network",
        "description": "Update a WireGuard network.",
        "operationId": "update_network_api_networks__network_id__put",
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Network Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WireGuardNetworkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireGuardNetworkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "networks",
          "networks"
        ],
        "summary": "Delete Network",
        "description": "Delete a WireGuard network.",
        "operationId": "delete_network_api_networks__network_id__delete",
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Network Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Network Api Networks  Network Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/networks": {
      "get": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Export Networks",
        "description": "Export all networks with their locations and devices.\n\nArgs:\n    exported_by: Who is performing the export\n    description: Optional description for the export\n    db: Database session\n\nReturns:\n    ExportData containing all networks, locations, and devices",
        "operationId": "export_networks_api_export_networks_get",
        "parameters": [
          {
            "name": "exported_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "Exported By"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Import Networks",
        "description": "Import networks from a JSON export file.",
        "operationId": "import_networks_api_export_networks_post",
        "parameters": [
          {
            "name": "imported_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "system",
              "title": "Imported By"
            }
          },
          {
            "name": "overwrite_existing",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overwrite Existing"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_networks_api_export_networks_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Import Networks Api Export Networks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export": {
      "post": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Export Networks From Request",
        "description": "Export networks from a request body.",
        "operationId": "export_networks_from_request_api_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/download": {
      "post": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Download Export",
        "description": "Download export data as a ZIP file.",
        "operationId": "download_export_api_export_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/networks/{network_id}/configs": {
      "get": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Export Network Configs",
        "description": "Export all device configurations for a network as a ZIP file.\n\nArgs:\n    network_id: ID of the network to export\n    db: Database session\n    format: Configuration format ('wg', 'json', 'mobile')\n    include_preshared_keys: Whether to include preshared keys\n    master_password: Master password for decrypting private keys\n\nReturns:\n    StreamingResponse with ZIP file containing all configurations\n\nRaises:\n    HTTPException: If network not found or export fails",
        "operationId": "export_network_configs_api_export_networks__network_id__configs_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Network Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Configuration format",
              "default": "wg",
              "title": "Format"
            },
            "description": "Configuration format"
          },
          {
            "name": "include_preshared_keys",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include preshared keys in configurations",
              "default": false,
              "title": "Include Preshared Keys"
            },
            "description": "Include preshared keys in configurations"
          },
          {
            "name": "master_password",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Master password for decrypting private keys",
              "title": "Master Password"
            },
            "description": "Master password for decrypting private keys"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit/export": {
      "get": {
        "tags": [
          "audit",
          "audit"
        ],
        "summary": "Export Audit Events",
        "description": "Export audit events with optional filtering.",
        "operationId": "export_audit_events_api_audit_export_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by network ID",
              "title": "Network Id"
            },
            "description": "Filter by network ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (ISO format)",
              "title": "Start Date"
            },
            "description": "Filter by start date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (ISO format)",
              "title": "End Date"
            },
            "description": "Filter by end date (ISO format)"
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor name",
              "title": "Actor"
            },
            "description": "Filter by actor name"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format (json or csv)",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format (json or csv)"
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include event details",
              "default": true,
              "title": "Include Details"
            },
            "description": "Include event details"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Export Audit Events Api Audit Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/export/download": {
      "get": {
        "tags": [
          "audit",
          "audit"
        ],
        "summary": "Download Audit Events",
        "description": "Download audit events as a file.",
        "operationId": "download_audit_events_api_audit_export_download_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by network ID",
              "title": "Network Id"
            },
            "description": "Filter by network ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (ISO format)",
              "title": "Start Date"
            },
            "description": "Filter by start date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (ISO format)",
              "title": "End Date"
            },
            "description": "Filter by end date (ISO format)"
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor name",
              "title": "Actor"
            },
            "description": "Filter by actor name"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format (json or csv)",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format (json or csv)"
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include event details",
              "default": true,
              "title": "Include Details"
            },
            "description": "Include event details"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WireGuardNetworkResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "network_cidr": {
            "type": "string",
            "title": "Network Cidr",
            "description": "Network CIDR notation (e.g., 10.0.0.0/24)"
          },
          "dns_servers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Servers",
            "description": "Comma-separated DNS servers (IP or domain)"
          },
          "mtu": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mtu",
            "description": "Maximum transmission unit (576-9000 bytes)"
          },
          "persistent_keepalive": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persistent Keepalive",
            "description": "Persistent keepalive interval in seconds (0-86400)"
          },
          "interface_properties": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface Properties",
            "description": "Additional WireGuard interface properties as key-value pairs"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "location_count": {
            "type": "integer",
            "title": "Location Count",
            "default": 0
          },
          "device_count": {
            "type": "integer",
            "title": "Device Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "network_cidr",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "WireGuardNetworkResponse",
        "description": "Schema for WireGuard network response."
      },
...
    }
  }
}